+++ TeamCode/src/main/java/IntoTheDeep2024/24OTOSAutoDrive.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package IntoTheDeep2024;
 
 import com.qualcomm.hardware.rev.RevHubOrientationOnRobot;
 import com.qualcomm.hardware.sparkfun.SparkFunOTOS;
