+++ TeamCode/src/main/java/IntoTheDeep2024/24RemoteControl.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package IntoTheDeep2024;
 
 // All the things that we use and borrow
 
@@ -13,9 +13,9 @@ import com.qualcomm.robotcore.util.RobotLog;
 
 import org.firstinspires.ftc.robotcore.external.navigation.CurrentUnit;
 
-@TeleOp(name="SurvivorCode", group="Linear OpMode")
+@TeleOp(name="24 Remote Control", group="Linear OpMode")
 @Disabled
-public class SurvivorCode extends LinearOpMode {
+public class RemoteControl extends LinearOpMode {
     // Initialize all variables for the program below:
     // This chunk controls our wheels
     private DcMotor leftFrontDrive = null;
@@ -27,6 +27,7 @@ public class SurvivorCode extends LinearOpMode {
     double leftBackPower = 0;
     double rightBackPower = 0;
     double max = 0;
+    boolean wheelClimb = false;
 
     // Collect joystick position data
     double axial = 0;
@@ -38,22 +39,23 @@ public class SurvivorCode extends LinearOpMode {
     final int VERTICAL_MIN = 0;
     final int VERTICAL_MAX = 1700;
     final int VERTICAL_MAX_VIPER = 1200;
+    final int VERTICAL_CLIMB_POSITION = 2300;
     final int VERTICAL_DEFAULT_SPEED = 2000;
     int verticalAdjustedMin = 0;
     int verticalPosition = VERTICAL_MIN;
 
     // This chunk controls our viper slide
     DcMotor viperSlide = null;
-    final int VIPER_MAX_WIDE = 2100;
-    final int VIPER_MAX_TALL = 3100;
+    final int VIPER_MAX_WIDE = 2000;
+    final int VIPER_MAX_TALL = 3300;
     final int VIPER_MIN = 0;
     int viperSlidePosition = 0;
 
     // This chunk controls our claw
     //Callie
     Servo claw = null;
-    final double CLAW_MIN = 0.2;        // Claw is closed
-    final double CLAW_MAX = 0.8;        // Claw is open
+    final double CLAW_MIN = 0.4;        // Claw is closed
+    final double CLAW_MAX = 0.93;        // Claw is open
     double claw_position = CLAW_MIN;
 
     Servo ascentStick = null;
@@ -70,6 +72,7 @@ public class SurvivorCode extends LinearOpMode {
 
         // Wait for the game to start (driver presses PLAY)
         telemetry.addData("Remote Control Ready", "press PLAY");
+        telemetry.addData("This code was last updated", "12/11/2024, 7:17 pm"); // Todo: Update this date when the code is updated
         telemetry.update();
         setAscentStick(ASCENT_MIN);
         waitForStart();
@@ -78,28 +81,62 @@ public class SurvivorCode extends LinearOpMode {
         // Run until the end of the match (driver presses STOP)
         while (opModeIsActive()) {
             double max;
+
             // Get input from the joysticks
             axial = -gamepad1.left_stick_y;
             lateral = gamepad1.left_stick_x;
             yaw = gamepad1.right_stick_x;
 
             setWheelPower();
-            leftFrontDrive.setPower(leftFrontPower);
-            rightFrontDrive.setPower(rightFrontPower);
-            leftBackDrive.setPower(leftBackPower);
-            rightBackDrive.setPower(rightBackPower);
+
+            if(!wheelClimb) {
+                // Send calculated power to wheels
+                leftFrontDrive.setPower(leftFrontPower);
+                rightFrontDrive.setPower(rightFrontPower);
+                leftBackDrive.setPower(leftBackPower);
+                rightBackDrive.setPower(rightBackPower);
+            }
 
             // Control the vertical - the rotation level of the arm
             verticalPosition = vertical.getCurrentPosition();
             verticalAdjustedMin = (int)(0.07*viperSlidePosition+VERTICAL_MIN); // 0.07 - If the viper is hits the ground, make this bigger. If it doesn't down far enough, make this smaller.
 
+            // Set vertical into initial climb position
+            if (gamepad1.dpad_up) {
+                // Hook onto the bar
+                setVertical(VERTICAL_CLIMB_POSITION, 2500);
+            }
+            // Active climb
+            else if (gamepad1.dpad_down) {
+                if (vertical.getCurrentPosition() > 100) {
+                    wheelClimb = true;
+                    vertical.setMode(DcMotor.RunMode.RUN_USING_ENCODER);
+                    vertical.setPower(-0.8);
+                    leftBackDrive.setPower(0.5);
+                    rightBackDrive.setPower(0.5);
+                }
+                else {
+                        leftBackDrive.setPower(0);
+                        rightBackDrive.setPower(0);
+                        wheelClimb = false;
+                        setVertical(VERTICAL_MIN, 1000);
+                }
+            }
+
             // If the right button is pressed AND it can safely raise further
-            if (gamepad1.dpad_right) {
+            else if (gamepad1.dpad_right && verticalPosition < VERTICAL_MAX) {
                 setVertical(Math.min(VERTICAL_MAX, verticalPosition + 50), 2000);
             }
             // If the left button is pressed AND it can safely lower without changing the viper
-            else if (gamepad1.dpad_left) {
-                setVertical( verticalPosition - 630);
+            else if (gamepad1.dpad_left && verticalPosition > VERTICAL_MAX_VIPER) {
+                setVertical(Math.max(VERTICAL_MAX_VIPER, verticalPosition - 50), 1500);
+            }
+            // If the left button is pressed AND it can safely lower further
+            else if (gamepad1.dpad_left && verticalPosition > verticalAdjustedMin) {
+                if (viperSlidePosition > VIPER_MAX_WIDE) {
+                    setViper(VIPER_MAX_WIDE, 1000);
+                }
+                setVertical(Math.max(verticalAdjustedMin, verticalPosition - 50),1000);
             }
 
             // Control the viper slide - how much it extends
@@ -117,21 +154,47 @@ public class SurvivorCode extends LinearOpMode {
                 viperSlide.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             }
             // If the right button is pressed AND it can safely retract further
-            else if (gamepad1.left_trigger > 0) {
-                viperSlide.setTargetPosition(viperSlidePosition - 200);
+            else if (gamepad1.left_trigger > 0 && viperSlidePosition > VIPER_MIN) {
+                viperSlide.setTargetPosition(Math.max(VIPER_MIN, viperSlidePosition - 200));
                 ((DcMotorEx) viperSlide).setVelocity(gamepad1.left_trigger*4000);
                 viperSlide.setMode(DcMotor.RunMode.RUN_TO_POSITION);
             }
 
             // Control the claw
             if (gamepad1.right_bumper && claw_position < CLAW_MAX) {
-                claw_position += 0.02;
+                claw_position += 0.3;
             }
             if (gamepad1.left_bumper && claw_position > CLAW_MIN) {
-                claw_position -= 0.015;
+                claw_position -= 0.15;
             }
             claw.setPosition(claw_position);
 
+            // Y/Triangle: High basket scoring position.
+            if (gamepad1.y) {
+                setVertical(VERTICAL_MAX, 3000);
+                setViper(VIPER_MAX_TALL, 2000);
+            }
+
+            // A/X button: Complete Retraction- Viper and vertical completely retracted and down
+            if (gamepad1.a) {
+                setViper(VIPER_MIN, 4000);
+                setVertical(VERTICAL_MIN, 700);
+            }
+
+            // X/Square: The viper slide is completely retracted but the vertical is in submersible position.
+            if (gamepad1.x) {
+                setVertical(355, 3000);
+                setViper(0, 1500);
+                wheelClimb = false;
+            }
+
+            // B/Circle: The vertical is in submersible position and the viper slide is all the way out.
+            if (gamepad1.b) {
+                setVertical(380, 1800);
+                setViper(1900, 2000);
+            }
+
+            // Show the elapsed game time and wheel power.
             printDataOnScreen();
         }
         claw.close();
@@ -168,6 +231,7 @@ public class SurvivorCode extends LinearOpMode {
         ascentStick.setDirection(Servo.Direction.REVERSE);
         ascentStick.setPosition(ASCENT_MIN);
     }
+
     private void setWheelPower(){
         leftFrontPower = (axial + lateral + yaw) / 2;
         rightFrontPower = (axial - lateral - yaw) / 2;
@@ -191,6 +255,7 @@ public class SurvivorCode extends LinearOpMode {
         sleep(1000);
         RobotLog.vv("Rockin' Robots", "Target: %4.2f, Current: %4.2f", target, ascentStick.getPosition());
     }
+
 
     public void setVertical(int height){
         setVertical(height, VERTICAL_DEFAULT_SPEED);
