+++ TeamCode/src/main/java/IntoTheDeep2024/24ServoAdjustments.java
@@ -1,4 +1,4 @@
-package org.firstinspires.ftc.teamcode;
+package IntoTheDeep2024;
 
 // All the things that we use and borrow
